[FIX #682] added static_casts to allow better cross plattform compilation

+++ src/core/common/math/linearAlgebra/WMatrixFixed.h
@@ -856,7 +856,7 @@ public:
     template< typename RHSValueT >
     void operator/=( const RHSValueT& rhs )
     {
-        operator=( ( *this ) / rhs );
+        operator=( ( *this ) / static_cast< double >( rhs ) );
     }
 
     /**
